networking: Rename pieces.rs->piece_announcement.rs
