+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/farm.rs
@@ -29,8 +29,8 @@ use subspace_farmer::{Identity, NodeClient, NodeRpcClient};
 use subspace_farmer_components::piece_caching::PieceMemoryCache;
 use subspace_networking::libp2p::identity::{ed25519, Keypair};
 use subspace_networking::utils::online_status_informer;
+use subspace_networking::utils::piece_announcement::announce_single_piece_index_with_backoff;
 use subspace_networking::utils::piece_provider::PieceProvider;
-use subspace_networking::utils::pieces::announce_single_piece_index_with_backoff;
 use subspace_proof_of_space::Table;
 use tokio::sync::broadcast;
 use tracing::{debug, error, info, warn};
