+++ crates/subspace-farmer/src/utils/farmer_piece_getter.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 use subspace_core_primitives::{Piece, PieceIndex};
 use subspace_farmer_components::plotting::{PieceGetter, PieceGetterRetryPolicy};
 use subspace_networking::utils::multihash::ToMultihash;
-use subspace_networking::utils::pieces::announce_single_piece_index_hash_with_backoff;
+use subspace_networking::utils::piece_announcement::announce_single_piece_index_hash_with_backoff;
 use subspace_networking::Node;
 use tracing::debug;
